Refactor poc to be anonlinkclient
